+++ backend/src/test/java/com/prograngers/backend/repository/notification/NotificationRepositoryTest.java
@@ -1,10 +1,10 @@
 package com.prograngers.backend.repository.notification;
 
 import com.prograngers.backend.entity.Notification;
-import com.prograngers.backend.entity.Review;
 import com.prograngers.backend.entity.comment.Comment;
 import com.prograngers.backend.entity.member.Member;
 import com.prograngers.backend.entity.problem.Problem;
+import com.prograngers.backend.entity.review.Review;
 import com.prograngers.backend.entity.solution.Solution;
 import com.prograngers.backend.repository.comment.CommentRepository;
 import com.prograngers.backend.repository.member.MemberRepository;
