+++ backend/src/test/java/com/prograngers/backend/support/fixture/ReviewFixture.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.support.fixture;
 
-import com.prograngers.backend.entity.Review;
-import com.prograngers.backend.entity.Review.ReviewBuilder;
+import com.prograngers.backend.entity.review.Review;
+import com.prograngers.backend.entity.review.Review.ReviewBuilder;
 import com.prograngers.backend.entity.member.Member;
 import com.prograngers.backend.entity.solution.Solution;
